Make block positions int in all places in the code

+++ presto-main/src/main/java/com/facebook/presto/block/uncompressed/UncompressedBlockCursor.java
@@ -14,7 +14,7 @@ public class UncompressedBlockCursor
     private final Slice slice;
     private final int positionCount;
 
-    private long position;
+    private int position;
     private int offset;
     private int size;
 
@@ -45,7 +45,7 @@ public class UncompressedBlockCursor
     @Override
     public int getRemainingPositions()
     {
-        return (int) (positionCount - (position + 1));
+        return positionCount - (position + 1);
     }
 
     @Override
@@ -87,7 +87,7 @@ public class UncompressedBlockCursor
     }
 
     @Override
-    public boolean advanceToPosition(long newPosition)
+    public boolean advanceToPosition(int newPosition)
     {
         if (newPosition >= positionCount) {
             position = positionCount;
@@ -123,14 +123,14 @@ public class UncompressedBlockCursor
     }
 
     @Override
-    public long getPosition()
+    public int getPosition()
     {
         checkReadablePosition();
         return position;
     }
 
     @Override
-    public long getCurrentValueEndPosition()
+    public int getCurrentValueEndPosition()
     {
         checkReadablePosition();
         return position;
